Create the basic functions in algorithm and genotype
Created the wave function measurement for all individuals and added comments for TODO things.
Created the initialization of parameters fo genotype and added the measure function for measuring the qRegister state.
Questions arised about the measurement vector type. If a bool vector is used, it's harder to use this algorithm in existing problems. The solution is 'Binary' but it's storage method is weird and unintuitive. Further investigation is needed.