Fix QIGA disaster bug and update graphs
The QIGA disaster operator called the initialize method which kept increasing the thetas array size, resulting in a exponential slowdown of the algorithm. Now that it's fixed, algorithm purrs like a kitten.

Updated the scripts and jars for graphs, removed extras.
Script can now be called with a 'only display results' mode (-d) and send notification on Slack upon completion (-n).
Those .log and .tmp files are used for the graph and should not be edited. They'll get removed and updated when starting a new comparison test.
BE SURE TO CHANGE THE HOOK ID AND CHANNEL NAME BEFORE UPLOADING TO SLACK.

Knapsack problem works for QIGA.
The classical knapsack however, doesn't work. It gets stuck on the first iteration.

Finally, added some knapsack testing files and reorganized the project.